moved to libstructure
